shellyupdater/requirements.txt
Django==2.2.11
paho-mqtt==1.5.0
pytz==2019.3
sqlparse==0.3.1
